from_slack/requirements.txt

Flask==1.0.2
google-api-python-client==1.12.8
google-cloud-pubsub==2.1.0
slack_sdk==3.1.0
slackclient==2.9.3
oauth2client
